requirements.txt
python-dotenv
openai==0.27.0
requests==2.25.1
google-cloud-speech
openai-whisper
pyttsx3